- support TW Chinese in update - in update, copy shutdown.sh only if it does not exist - move check smc in update.sh to setup.sh - update changelog

+++ update.sh
@@ -8,6 +8,58 @@ function valid_day() {
 	fi
 }
 
+function get_changelog { # get the latest changelog
+	if [[ -z $1 ]]; then
+		changelog=$(curl -sSL https://raw.githubusercontent.com/js4jiang5/BatteryOptimizer_for_MAC/main/CHANGELOG | sed s:\":'\\"':g 2>&1)
+	else
+		changelog=$(curl -sSL https://raw.githubusercontent.com/js4jiang5/BatteryOptimizer_for_MAC/main/$1 | sed s:\":'\\"':g 2>&1)
+	fi
+
+    n_lines=0
+	while read -r "line"; do
+		num=$(echo $line | tr '.' ' '| tr 'v' ' ') # extract number parts
+		is_version=true
+		n_num=0
+		for var in $num; do
+			if ! [[ "$var" =~ ^[0-9]+$ ]]; then
+				is_version=false
+				break
+			else
+				n_num=$((n_num+1))
+			fi
+		done
+		if [[ $line =~ "." ]] && [[ $line =~ "v" ]] && $is_version && [[ $n_num == 3 ]] && [[ $n_lines > 0 ]]; then
+			is_version=true
+		else
+			is_version=false
+		fi
+
+        if $is_version; then # found the start of 2nd version
+			break
+		fi
+        n_lines=$((n_lines+1))
+    done <<< "$changelog"
+    echo -e "$changelog" | awk 'NR>=2 && NR<='$n_lines
+}
+
+function format00() {
+	value=$1
+	if [ $value -lt 10 ]; then
+		value=0$(echo $value | tr -d '0')
+		if [ "$value" == "0" ]; then
+			value="00"
+		fi
+	fi
+	echo $value
+}
+
+function version_number { # get number part of version for comparison
+	version=$1
+	num=$(echo $version | tr '.' ' '| tr 'v' ' ')
+	v1=$(echo $num | awk '{print $1}'); v2=$(echo $num | awk '{print $2}'); v3=$(echo $num | awk '{print $3}');
+	echo $(format00 $v1)$(format00 $v2)$(format00 $v3)
+}
+
 # Force-set path to include sbin
 PATH="$PATH:/usr/sbin"
 
@@ -16,12 +68,29 @@ tempfolder=~/.battery-tmp
 binfolder=/usr/local/bin
 configfolder=$HOME/.battery
 batteryfolder="$tempfolder/battery"
+language_file=$configfolder/language.code
 mkdir -p $batteryfolder
+
+lang=$(defaults read -g AppleLocale)
+if test -f $language_file; then
+	language=$(cat "$language_file" 2>/dev/null)
+	if [[ "$language" == "tw" ]]; then
+		is_TW=true
+	else
+		is_TW=false
+	fi
+else
+	if [[ $lang =~ "zh_TW" ]]; then
+		is_TW=true
+	else
+		is_TW=false
+	fi
+fi
 
 echo -e "🔋 Starting battery update\n"
 
+version_local=$(echo $(battery version)) #update informed version first
 # Write battery function as executable
-
 echo "[ 1 ] Downloading latest battery version"
 update_branch="main"
 in_zip_folder_name="BatteryOptimizer_for_MAC-$update_branch"
@@ -39,42 +108,29 @@ chown $USER $binfolder/battery
 chmod 755 $binfolder/battery
 chmod u+x $binfolder/battery
 
-if [[ $(smc -k BCLM -r) == *"no data"* ]]; then # power limit during shutdown only required for Apple CPU Macbook
-	echo "[ 3 ] Setup for power limit when Macs shutdown"
-	sudo cp $batteryfolder/dist/.reboot $HOME/.reboot
-	sudo cp $batteryfolder/dist/.shutdown $HOME/.shutdown
-	sudo cp $batteryfolder/dist/shutdown.sh $binfolder/shutdown.sh
-	sudo cp $batteryfolder/dist/battery_shutdown.plist $HOME/Library/LaunchAgents/battery_shutdown.plist
-	launchctl enable "gui/$(id -u $USER)/com.battery_shutdown.app"
-	launchctl unload "$HOME/Library/LaunchAgents/battery_shutdown.plist" 2> /dev/null
-	launchctl load "$HOME/Library/LaunchAgents/battery_shutdown.plist" 2> /dev/null
-	sudo chown $USER $HOME/.reboot
-	sudo chmod 755 $HOME/.reboot
-	sudo chmod u+x $HOME/.reboot
-	sudo chown $USER $HOME/.shutdown
-	sudo chmod 755 $HOME/.shutdown
-	sudo chmod u+x $HOME/.shutdown
-	sudo chown $USER $binfolder/shutdown.sh
-	sudo chmod 755 $binfolder/shutdown.sh
-	sudo chmod u+x $binfolder/shutdown.sh
-fi
-
-# Check if smc works
-check_smc=$(smc 2>&1)
-if [[ $check_smc =~ " Bad " ]] || [[ $check_smc =~ " bad " ]] ; then # current is not a right version
-	sudo cp $batteryfolder/dist/smc_intel $binfolder/smc
-	sudo chown $USER $binfolder/smc
-	sudo chmod 755 $binfolder/smc
-	sudo chmod +x $binfolder/smc
-	# check again
-	check_smc=$(smc 2>&1)
-	if [[ $check_smc =~ " Bad " ]] || [[ $check_smc =~ " bad " ]] ; then # current is not a right version
-		echo "Error: BatteryOptimizer seems not compatible with your MAC yet"
-		exit
+if ! test -f $binfolder/shutdown.sh; then # check if shutdown.sh already exist, to be removed at the beginning of 2025
+	if [[ $(smc -k BCLM -r) == *"no data"* ]]; then # power limit during shutdown only required for Apple CPU Macbook
+		echo "[ 3 ] Setup for power limit when Macs shutdown"
+		sudo cp $batteryfolder/dist/.reboot $HOME/.reboot
+		sudo cp $batteryfolder/dist/.shutdown $HOME/.shutdown
+		sudo cp $batteryfolder/dist/shutdown.sh $binfolder/shutdown.sh
+		sudo cp $batteryfolder/dist/battery_shutdown.plist $HOME/Library/LaunchAgents/battery_shutdown.plist
+		launchctl enable "gui/$(id -u $USER)/com.battery_shutdown.app"
+		launchctl unload "$HOME/Library/LaunchAgents/battery_shutdown.plist" 2> /dev/null
+		launchctl load "$HOME/Library/LaunchAgents/battery_shutdown.plist" 2> /dev/null
+		sudo chown $USER $HOME/.reboot
+		sudo chmod 755 $HOME/.reboot
+		sudo chmod u+x $HOME/.reboot
+		sudo chown $USER $HOME/.shutdown
+		sudo chmod 755 $HOME/.shutdown
+		sudo chmod u+x $HOME/.shutdown
+		sudo chown $USER $binfolder/shutdown.sh
+		sudo chmod 755 $binfolder/shutdown.sh
+		sudo chmod u+x $binfolder/shutdown.sh
 	fi
 fi
 
-# correct the schedule plist if it is incorrect due to the bug
+# correct the schedule plist if it is incorrect due to the bug, to be removed at the beginning of 2025
 schedule_tracker_file="$configfolder/calibrate_schedule"
 enable_exist="$(launchctl print gui/$(id -u $USER) | grep "=> enabled")"
 if [[ $enable_exist ]]; then # new version that replace => false with => enabled
@@ -138,7 +194,9 @@ if test -f $schedule_tracker_file && [[ $schedule_enabled =~ "enabled" ]]; then
 fi
 
 echo "[ 3 ] Setting up visudo declarations"
-sudo $batteryfolder/battery.sh visudo $USER
+if [[ $(version_number $version_local) < $(version_number "v2.0.9") ]]; then
+	sudo $batteryfolder/battery.sh visudo $USER
+fi
 
 # Remove tempfiles
 cd
@@ -155,9 +213,23 @@ echo "$version" > $informed_version_file
 
 battery maintain stop >> /dev/null
 sleep 1
-battery maintain_synchronous recover >> $HOME/.battery/battery.log &
-battery create_daemon >> /dev/null
-battery schedule enable >> /dev/null
-battery status 
+pkill -f "$binfolder/battery.*"
 
-echo -e "You're running the latest version $version now.\n"
+
+if [[ $(version_number $version_local) > $(version_number "v2.0.8") ]]; then
+	battery maintain recover
+else # to be removed at the beginning of 2025
+	battery maintain_synchronous recover >> $HOME/.battery/battery.log &
+	battery create_daemon >> /dev/null
+	battery schedule enable >> /dev/null
+	battery status 
+fi
+
+button_empty="                                                                                                                                                    "
+if $is_TW; then
+	changelog=$(get_changelog CHANGELOG_TW)
+	answer="$(osascript -e 'display dialog "'"已更新至 $version, 更新內容如下\n\n$changelog"'" buttons {"'"$button_empty"'", "完成"} default button 2 with icon note with title "BatteryOptimizer for MAC"' -e 'button returned of result')"
+else
+	changelog=$(get_changelog CHANGELOG)
+	answer="$(osascript -e 'display dialog "'"Update to $version completed, changes include\n\n$changelog"'" buttons {"'"$button_empty"'", "Finish"} default button 2 with icon note with title "BatteryOptimizer for MAC"' -e 'button returned of result')"
+fi
